requirements.txt
python-dotenv
numpy
pandas
certifi
scikit-learn
pyam
dill 
mlflow
dagshub
fastapi
uvicorn
pymongo[srv]

## -e .